fixed code after first review

+++ src/main/java/core/basesyntax/Circle.java
@@ -3,7 +3,7 @@ package core.basesyntax;
 public class Circle extends Figure {
     private final double radius;
 
-    public Circle(String color, int radius) {
+    public Circle(String color, double radius) {
         super(color);
         this.radius = radius;
     }
